+++ CHANGELOG.md
@@ -38,10 +38,12 @@ Changelog
         ```
 - **[IR]**: Rework assisted inject bindings to be encapsulated by their consuming assisted factory bindings in graph validation.
     - This ensures these classes can't accidentally participate in `SwitchingProvider`s or valid cycle breaking with `DelegateFactory`, as both of those require `Provider` types and assisted-inject types' factories don't implement `Provider`.
+- **[Gradle]**: Avoid deprecated `KotlinCompilation.implementationConfigurationName` API.
 - Fold `2.3.20-dev-7791` compat into `2.3.20-Beta2` compat, meaning the former is no longer tested on CI.
 - Test Kotlin `2.3.20-Beta2`.
 - Test Kotlin `2.3.10`.
 - Drop testing of Kotlin `2.3.10-RC`.
+- Metro now _compiles_ against Kotlin `2.3.0`. This allows it to test `2.4.0` builds, but is still compatible down to Kotlin `2.2.20`. Metro's runtime artifacts also explicitly have their language version still set to `2.2` (and `2.0` for the Gradle plugin).
 
 ### Contributors